feat: simplify `jest` integration
Makes reporter default export for easier integration with `jest`.
Moves `getCrapReport` API out of the index file for clarity.

Changes default behaviour for reporter to write both JSON and HTML reports by default,
so it is usable without any configuration, again for easier integration with `jest`.
Adds input validation for reporter options.
