+++ README.md
@@ -1,6 +1,6 @@
 # CRAP Score
 
-Calculate and visualize the CRAP score of a JS/TS project using the provided API or CLI.
+Calculate and visualize the CRAP score of a JS/TS project using the provided `jest` integration, CLI, or API.
 
 ## Example
 
@@ -18,33 +18,38 @@ Combining complexity and coverage information, the CRAP score gives you insight 
 
 ## How to Use
 
-### Test Reporter
+### Jest Reporter
 
-#### Jest
+Add `crap-score` as [test reporter](https://jestjs.io/docs/configuration#reporters-arraymodulename--modulename-options) to jest.
+When `jest` is run with coverage enabled, this will also generate a CRAP report. Example for `jest.config.json`:
 
-Add `crap-score` as [test reporter to jest](https://jestjs.io/docs/configuration#reporters-arraymodulename--modulename-options).
-This only works when `jest` is run with `--coverage`. Example for `jest.config.json`:
+```json
+"reporters": ["default", "crap-score"],
+```
+
+The reporter also accepts options, for example:
 
 ```json
 "reporters": [
     "default",
     [
-        "crap-score/dist/test-reporter/index.js",
+        "crap-score",
         {
-            "htmlReportDir": "./crap-report/html"
+            "jsonReportFile": "crap.json",
         }
     ]
 ]
 ```
 
-A typed interface for the reporter options is available under `import type { ReporterOptions } from "crap-score";`.
+A typed interface for the reporter options is available via `import type { ReporterOptions } from "crap-score";`.
+
 Available options:
 
-| Option         | Description                                                                                              |
-| -------------- | -------------------------------------------------------------------------------------------------------- |
-| jsonReportFile | Specifies path where the JSON report will be written to. If undefined, no report is written to the disk. |
-| htmlReportDir  | Specifies path where the HTML report will be written to. If undefined, no report is written to the disk. |
-| log            | Changes log behaviour. `"silent"` suppresses all logs. `"debug"` prints additional logs.                 |
+| Option         | Description                                                                                                                               |
+| -------------- | ----------------------------------------------------------------------------------------------------------------------------------------- |
+| jsonReportFile | Specifies path where the JSON report will be written to. Defaults to `crap-report/crap-report.json`. Pass `false` to disable JSON report. |
+| htmlReportDir  | Specifies path where the HTML report will be written to. Defaults to `crap-report/html/`. Pass `false` to disable HTML report.            |
+| log            | Changes log behaviour. `"silent"` suppresses all logs. `"debug"` prints additional logs.                                                  |
 
 ### CLI
